Add kernelci-storage role
When callback/boot is used, kernelci-celery try to store boot result in
/var/www/images.

This patch add handling of this directory by ansible.
Since this directory us used also by the "storage" vhost, and that
storage could be seen as a part of backend, it is easier to add a
kernelci-storage role.