Add kci_app_name and kci_app_port variables
This commit adds kci_app_name to differentiate between instances of
kernelci-backend. This can be useful if there is a need to deploy
several independent kernelci-backend installations on a single machine.
It adds an application name to service and config files so they don't
get overwritten or confused during deployment.

kci_app_port variable is required now to specify the port on which
kernelci-backend is running.

NOTE: Ansible playbook doesn't configure and at the moment can't rename
db name

Signed-off-by: Michal Galka <michal.galka@collabora.com>