Fix SELinux status detection
Checks if ansible_selinux variable is defined and ansible_selinux.status
is explicitly marked as 'enabled' before including selinux related
steps.
Before it used to check if ansible_selinux is defined and is not
false, but this fails for example when the selinux python module is not
installed. In such case ansible_selinux remains defined and
ansible_selinux.status stores an error message, which is lead to false
positive errors.

Signed-off-by: Michal Galka <michal.galka@collabora.com>